Save recent efforts approaching difficulties in arglist detection
Hard to detect lamdba formal parameters. Might need to only do for
ones preceded by identifier. Also, it seems it should be treated as
continued statement if there is no comma, which comes later.
arglist-cont might at least be easier. But it seems we need to cook up
how to parse formal parameters. Maybe similarly to how we parse
statements.