Add Travis-CI and Cask
We chose EVM for the emacs images in travis. We avoid the docker emacs
images for now, as it made it less straight-forward to setup local
travis-ci testing, via the travis-ci docker image, and then the steps
required to nest a docker emacs inside.

The current .travis.yml file was adapted from EVM's README and elpy's
.travis.yml.

The use of Cask is motivated by the external dependencies, such as
dash and s.