Adds failing test against dart dev
This test succeeds on stable branch of dart, and fails on development
branch of dart.

This test is fragile, relying on events that are not completely
understood by its author. Especially, the process seems to return at
least twice, and we have to limit the time for the third return.

Another instance of difficulty in testing of asynchronous processes
can be seen at
https://lists.gnu.org/archive/html/emacs-devel/2013-02/msg00506.html.