Allow type parameters in function signatures
Introduces changes to fontify function delarations with type
parameters in the function signature, like in the following.

    T first<T>(List<T> ts) {
      ...
    }

To make this easier, we'd like `up-list` to work inside angle
brackets, so we've modified the syntax of angle brackets from symbols
to parens.

This was failing on short words, like `<S>`, so to make _that_ task
easier, by using `word-start`, we've changed the syntax of $ and _ to
words.

That last choice, may cause issues unforeseen. If it doesn't, it might
help to clean up code elsewhere.