Treat angle brackets as puncutation syntax instead of symbol syntax
When using for example dabbrev-expand completion, counting angle
brackets as symbols leads to undesirable behavior.

Eventually, we might treat it as paren syntax, but we'd then have to
take care to not treat it as paren syntax when they appear as
operators, like in `2 < 3`.

Treating as punctuation syntax should be a good compromise and
improvement for now.