Restore copyright information and discuss previous implementation
To a not fully determined extend, this implementation derives from an
earlier implementation by Natalie Weizenbaum, for which the copyright
holder is Google.

In the discussion attached to [1], Brady Trainor implicitly repeats
his claim, that no code from that earlier implementation remains.
This seems plausible, but, as mentioned in that discussion, would have
to be studied in more detail, to be absolutely certain, which is why I
am restoring Google's copyright notice.

Also state that Brady Trainor is the author of *this* implementation,
but remove him as a maintainer, because he has handed this package
over to the Emacsorphanage, where Jen-Chieh Shen has picked up the
maintainer role.

[1] https://github.com/emacsorphanage/dart-mode/commit/69d7c23be7518bcb004860120ba3ea2a7dcfb6de.